+++ common/src/main/java/earth/terrarium/cadmus/common/network/NetworkHandler.java
@@ -13,7 +13,7 @@ public class NetworkHandler {
     public static final Network CHANNEL = new Network(new ResourceLocation(Cadmus.MOD_ID, "main"), 1, true);
 
     public static void init() {
-        CHANNEL.register(ServerboundSendSilentChatCommandPacket.TYPE);
+        CHANNEL.register(ServerboundSendClaimChatCommandPacket.TYPE);
         CHANNEL.register(ClientboundSyncClaimsPacket.TYPE);
         CHANNEL.register(ClientboundAddClaimPacket.TYPE);
         CHANNEL.register(ClientboundAddClaimsPacket.TYPE);